use pubsub to stop running spider

+++ dragline/listener.py
@@ -1,75 +1,45 @@
-
 import thread
-
-
 from redisds import Queue
 from redis import StrictRedis
 import os
 from subprocess import Popen
 
 
-re=StrictRedis()
-start=Queue(name="start",namespace="dragd")
-start.put("1")
+process = None
+run_id = None
 
-stop=Queue(name="stop",namespace="dragd")
 
-global process
 def listen_start():
+    global run_id, process
+    start = Queue(name="start", namespace="dragd", db=1)
     try:
-
         while True:
-
-
-            run_id=start.get()
-
-            global process
-
+            run_id = start.get()
             FNULL = open(os.devnull, 'w')
-            comm=["python","start.py",run_id]
-
-            print comm
-            process=Popen(comm)
-
-
-
-            #process.wait()
-
-
-
-
-
-
-
+            comm = ["python", "start.py", run_id]
+            process = Popen(comm)
+            process.wait()
+            run_id = None
     except Exception as e:
         print "error in start"
 
 
-
 def listen_stop():
-
-
+    global process, run_id
+    redisclient = StrictRedis(db=1)
+    pubsub = redisclient.pubsub()
+    pubsub.subscribe("dragd:stop")
     while True:
         try:
-            global process
-
-
-            run_id = stop.get()
-
-
-            process.terminate()
+            for i in pubsub.listen():
+                if i['data'] == run_id:
+                    process.terminate()
+                    print run_id, "stopped"
         except Exception as e:
             print "error"
-
+            print traceback.format_exc()
 
 
 if __name__ == "__main__":
-    thread.start_new_thread(listen_stop,())
+    thread.start_new_thread(listen_stop, ())
     listen_start()
-
-
-
-
-
-
-
